+++ frontend/src/components/CalendarMobile/CalendarMobile.js
@@ -37,7 +37,7 @@ function eventStyleGetter(event) {
         color: "white",
         border: "0.15em solid lightgrey",
         paddingLeft: "0.5em",
-        fontWeight: "300",
+        fontWeight: "400",
         fontSize: "0.9em",
     };
     return {
@@ -57,7 +57,7 @@ class CalendarMobile extends Component {
                     events={this.props.events}
                     startAccessor="start"
                     endAccessor="end"
-                    defaultView="work_week"
+                    defaultView="day"
                     views={["work_week", "day"]}
                     min={new Date(1970, 1, 1, 8, 0, 0)}
                     max={new Date(1970, 1, 1, 20, 0, 0)}